+++ docs/HISTORY.rst
@@ -1,11 +1,15 @@
 Changelog
 =========
 
-0.1.4 (unreleased)
+0.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Move resources to bundle
+  [cekk]
+- Fix code-quality
+  [cekk]
+- Add travis config
+  [cekk]
 
 0.1.3 (2017-07-03)
 ------------------